Code/models/mnE_c2f2_trip.py
'''
Copyright (C) 2019-2020, Authors of ECCV2020 #2274 "Adversarial Ranking Attack and Defense"
Copyright (C) 2019-2020, Mo Zhou <cdluminate@gmail.com>

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
'''
import os
import torch as th
import collections
import math
import yaml
from . import datasets
from . import faE_c2f2_trip

class Model(faE_c2f2_trip.Model):
    """
    LeNet-like convolutional neural network for embedding
    """
    def getloader(self, kind:str='train', batchsize:int=1):
        '''
        get corresponding dataloaders
        '''
        config = yaml.load(open('config.yml', 'r').read(),
                Loader=yaml.SafeLoader)
        if kind == 'train':
            return datasets.fashion.get_loader(
                    os.path.expanduser(config['mnist']['path']),
                    batchsize, 'train')
        else:
            return datasets.fashion.get_loader(
                    os.path.expanduser(config['mnist']['path']),
                    batchsize, 't10k')
