Code/requirements.txt
numpy
pandas
pylab
scipy
seaborn
termcolor
torch (>= 1.0.0)
torchvision
tqdm
PyYAML
